extracting abstract output class
